Moved componenets and styles
