+++ src/pages/explore-options.tsx
@@ -1,6 +1,6 @@
 import React from "react";
-import Layout from "../components/Layout";
-import OptionViewContainer from "../frontend/componenets/OptionViewContainer";
+import Layout from "../frontend/components/Layout";
+import OptionViewContainer from "../frontend/components/OptionViewContainer";
 
 function ExploreOptions()
 {
