+++ src/styles/components/OptionDetailsView.module.scss
