+++ src/styles/components/OptionListItemView.module.scss
