+++ src/styles/components/OptionViewContainer.module.scss
