+++ src/styles/pages/404.module.scss
